@@ -1,12 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+    <DockerfileContext>..\..\..</DockerfileContext>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.8.1" />
@@ -15,8 +13,7 @@
     <PackageReference Include="Castle.Core" Version="4.3.1" />
     <PackageReference Include="DotNet.Standard.NParsing.SQLServer" Version="1.0.3" />
     <PackageReference Include="DotNet.Standard.NSmart" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -24,12 +21,6 @@
     <ProjectReference Include="..\DotNet.Core.Demo.Models\DotNet.Core.Demo.Models.csproj" />
     <ProjectReference Include="..\DotNet.Core.Demo.Services\DotNet.Core.Demo.Services.csproj" />
     <ProjectReference Include="..\DotNet.Core.Demo.Utilities\DotNet.Core.Demo.Utilities.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="Dockerfile">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
   </ItemGroup>
 
 </Project>
