@@ -1,7 +1,5 @@
-﻿using System;
 using System.Reflection;
 using Autofac;
-using Autofac.Extensions.DependencyInjection;
 using Autofac.Extras.DynamicProxy;
 using DotNet.Core.Demo.IServices;
 using DotNet.Core.Demo.Models;
@@ -9,9 +7,9 @@ using DotNet.Core.Demo.WebAPI.Filters;
 using DotNet.Standard.NSmart;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json;
 
 namespace DotNet.Core.Demo.WebAPI
@@ -26,43 +24,48 @@ namespace DotNet.Core.Demo.WebAPI
         public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
-        public IServiceProvider ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(IServiceCollection services)
         {
-            //依赖注入的配置
             services.AddMvc(option =>
             {
                 option.Filters.Add(typeof(CustomerActionFilterAttribute));
                 option.Filters.Add(typeof(CustomerExceptionFilterAttribute));
-            }).AddJsonOptions(option =>
+            }).AddNewtonsoftJson(option =>
             {
                 option.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
                 option.SerializerSettings.DateFormatString = "yyyy-MM-dd HH:mm:ss.fff";
                 option.SerializerSettings.Converters.Add(new DoModelConverter());
-            }).AddControllersAsServices().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-            var builder = new ContainerBuilder();
-            builder.Populate(services);
-            //注册拦截器类
+            });
+            services.AddControllersWithViews().AddControllersAsServices();
+        }
+
+        public void ConfigureContainer(ContainerBuilder builder)
+        {
+            //ע����������
             builder.RegisterType<ServiceInterceptor>();
             builder.RegisterAssemblyTypes(Assembly.Load("DotNet.Core.Demo.Services"))
                 .Where(type => typeof(IBaseService).IsAssignableFrom(type) && !type.IsAbstract)
-                .AsImplementedInterfaces() //根据名称约定，实现服务接口和服务实现的依赖
-                .InstancePerLifetimeScope() //生命周期
+                .AsImplementedInterfaces() //��������Լ����ʵ�ַ���ӿںͷ���ʵ�ֵ�����
+                .InstancePerLifetimeScope() //��������
                 .PropertiesAutowired()
-                .EnableInterfaceInterceptors() //启用拦截器
+                .EnableInterfaceInterceptors() //����������
                 .InterceptedBy(typeof(ServiceInterceptor));
             builder.RegisterAssemblyTypes(Assembly.GetEntryAssembly()).PropertiesAutowired();
-            var container = builder.Build();
-            return new AutofacServiceProvider(container);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
-            app.UseMvc();
+            app.UseRouting();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
